// src/gen/src/proto/member/v1/member_connect.ts
// @generated by protoc-gen-connect-es v1.6.1 with parameter "target=ts"
// @generated from file src/proto/member/v1/member.proto (package member.v1, syntax proto3)
/* eslint-disable */
// @ts-nocheck

import { CreateMemberRequest, CreateMemberResponse, DeleteMemberRequest, DeleteMemberResponse, GetMemberRequest, GetMemberResponse, ListMembersRequest, ListMembersResponse, UpdateMemberRequest, UpdateMemberResponse } from "./member_pb.js";
import { MethodKind } from "@bufbuild/protobuf";

/**
 * @generated from service member.v1.MemberService
 */
export const MemberService = {
  typeName: "member.v1.MemberService",
  methods: {
    /**
     * @generated from rpc member.v1.MemberService.CreateMember
     */
    createMember: {
      name: "CreateMember",
      I: CreateMemberRequest,
      O: CreateMemberResponse,
      kind: MethodKind.Unary,
    },
    /**
     * @generated from rpc member.v1.MemberService.GetMember
     */
    getMember: {
      name: "GetMember",
      I: GetMemberRequest,
      O: GetMemberResponse,
      kind: MethodKind.Unary,
    },
    /**
     * @generated from rpc member.v1.MemberService.UpdateMember
     */
    updateMember: {
      name: "UpdateMember",
      I: UpdateMemberRequest,
      O: UpdateMemberResponse,
      kind: MethodKind.Unary,
    },
    /**
     * @generated from rpc member.v1.MemberService.DeleteMember
     */
    deleteMember: {
      name: "DeleteMember",
      I: DeleteMemberRequest,
      O: DeleteMemberResponse,
      kind: MethodKind.Unary,
    },
    /**
     * @generated from rpc member.v1.MemberService.ListMembers
     */
    listMembers: {
      name: "ListMembers",
      I: ListMembersRequest,
      O: ListMembersResponse,
      kind: MethodKind.Unary,
    },
  }
} as const;